Unify untar code.
